+++ app/models.py
@@ -1,9 +1,12 @@
 import uuid
 import contextvars
 from hashlib import md5
+import warnings
 
 import numpy as np
 import sqlalchemy as sa
+import traceback
+import requests
 from slugify import slugify
 from sqlalchemy import func
 from sqlalchemy.dialects.postgresql import JSONB
@@ -14,9 +17,42 @@ from sqlalchemy_utils import EmailType, PhoneNumberType
 
 from .custom_exceptions import AccessError
 from .json_util import to_json
+from .env import load_env
 
 session_context_id = contextvars.ContextVar('request_id', default=None)
 DBSession = scoped_session(sessionmaker(), scopefunc=session_context_id.get)
+
+env, cfg = load_env()
+strict = cfg['security']['strict']
+use_webhook = cfg['security']['slack']['enabled']
+webhook_url = cfg['security']['slack']['url']
+
+
+def verify(mode, row, accessor):
+    """Verifies that User or Token `accessor` can access `Base` row
+    in mode `mode` (can be create, read, update, or delete)."""
+
+    if not row.is_accessible_by(accessor, mode=mode):
+        tb = ''.join(traceback.extract_stack().format()[-3:-1])
+        tb = f'```{tb}```'
+        err_msg = (
+            f'Insufficient permissions for operation '
+            f'"{type(accessor).__name__} {accessor.id} '
+            f'{mode} {type(row).__name__} {row.id}". Original traceback: {tb}'
+        )
+        if use_webhook:
+            try:
+                requests.post(webhook_url, json={'text': err_msg})
+            except requests.HTTPError as e:
+                post_fail_warn_msg = f'Encountered HTTPError "{e.args[0]}" ' \
+                                     f'attempting to post AccessError "{err_msg}"' \
+                                     f'to {webhook_url}.'
+                warnings.warn(post_fail_warn_msg)
+        else:
+            warnings.warn(err_msg)
+        if strict:
+            raise AccessError(err_msg)
+
 
 # https://docs.sqlalchemy.org/en/13/dialects/postgresql.html#psycopg2-fast-execution-helpers
 # executemany_values_page_size arguments control how many parameter sets
